backend/requirements.txt
# Web Framework
flask==3.0.3
flask-cors==4.0.1

# Environment Configuration
python-dotenv==1.0.1

# Vector Search and Embeddings
# faiss-cpu 1.7.4 is compatible with Python 3.11/3.12
# numpy 1.26.4 works well with faiss-cpu 1.7.4
faiss-cpu==1.7.4
numpy>=1.24.0,<1.27.0
nomic==3.6.0

# HTTP Client for direct API calls
requests==2.31.0

# AI/ML Services
google-generativeai==0.8.3

# Production Server (optional, only needed for deployment)
gunicorn==23.0.0 